Add floating button to the android application.
Closes: https://github.com/singh264/expenses/issues/54.
Author: Amarpreet Singh amarpreet1997@gmail.com.
